Add a mechanism to easily embed local files when preprocessing an mpp file
Basically, using the following structure:
```
"mpp-include-base64": [
    {"id": "README", "path": "README.md", "sha256": "301684a4b434fe51f287cf74360634b1800defec92744ce59400e4a6f36846d8"},
    {"id": "rhel8-rpi4-mpp", "path": "rhel8-rpi4.mpp.json"}
],
```
one can define a list of files that one wants included in the template.
This `id` contains the name given to the variable corresponding to this
file.
`path` contains a full path to the file one wants to embed, relative to
the template mpp file.
`sha256` or `sha512` are optional fields that can be used to ensure that
the file being embedded in the template is the expected on. If provided
these fields will be checked against the sha256/sha512 computed upon
reading the file.

You can then install the file using the following structure:
```
{
    "type": "org.osbuild.copy",
    "inputs": {
        "inlinefile": {
            "type": "org.osbuild.files",
            "origin": "org.osbuild.source",
            "references": {"mpp-format-references": [
            	{"id": "README", "args": {"foo": "bar", "baz": "boum"}},
            	{"id": "rhel8-rpi4-mpp"}
            ]}
        }
    },
    "options": {
        "paths": [
            {
                "from": { "mpp-format-string": "input://inlinefile/{foo.digest}"},
                "to": "tree:///usr/bin/start.sh"
            }
        ]
    }
}
```

Where `mpp-format-references` will generate the following structure
based on the `README` and `rhel8-rpi4-mpp` variables (defined by the
`id` in  `mpp-include-base64` above):
```
references": {
    "sha256:301684a4b434fe51f287cf74360634b1800defec92744ce59400e4a6f36846d8": {
      "foo": "bar",
      "baz": "boum"
    },
    "sha256:41d3afc0adb06d96397e082cb5192416adf80fd70ed767953c4b36ac7abc655a": {}
}
```

And `{ "mpp-format-string": "input://inlinefile/{foo.digest}"},` will
generate the actual content of the file as base64 as supported by osbuild
e.g.:
```
"paths": [
    {
        "from": "input://inlinefile/sha256:301684a4b434fe51f287cf74360634b1800defec92744ce59400e4a6f36846d8"
        "to": "tree:///usr/bin/start.sh"
    }
]
```

Finally, the code will automatically insert the "org.osbuild.inline"
sources for the file that are included.
This will result in a structure such as:
```
  "org.osbuild.inline": {
    "items": {
      "sha256:301684a4b434fe51f287cf74360634b1800defec92744ce59400e4a6f36846d8": {
        "encoding": "base64",
        "data": "VGhpcyBkaXJlY3RvcnkgY29udGFpbnMgc29tZSBzYW1wbGUgb3NidWlsZ...
      }
    }
  }
```

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>