stages/users: create home before running usermod with move
When building ostree upgrade commits (commits with a parent), we inherit
the passwd and group files from the parent commit.  While users might
exist in the parent and specified again in the new commit, the home
directory wont exist in the tree.

Before running `usermod`, create the old home directory (and chown it),
so that
1. potentially running `usermod --home <newhome> --move-home` doesn't
   fail (<oldhome> not found),
2. the new ssh key can be created in the home directory of an existing
   ostree user.