stages/rpm: support systemd-first-boot
The protocol described in  `machite-id(5)` describes how `ConditionFirstBoot` is
triggered in case `/etc/machine-id` does not exist during boot. Apart from
triggering this condition, the behaviour is the same as when `/etc/machine-id` is
empty.

When `first-boot` is set to `true`, /etc/machine-id is not created unless it existed. If `first-boot` is false (or unset) an empty `/etc/machine-id` is created
as before.